fix: move fetchTestimonials to Testimonials folder and add console.error
